requirements.txt
langchain==0.0.131
streamlit-option-menu==0.3.2
streamlit-toggle-switch==1.0.2
streamlit-chat==0.0.2.2
streamlit-authenticator==0.2.1
openai==0.27.2
gspread==5.7.2
oauth2client==4.1.3
pandas==1.5.3
Pillow==9.4.0
gspread-dataframe==3.3.0
requests==2.28.2